rf: create new ``header`` module of interfaces
Contains ``CopyXForm|CopyHeader|ValidateImage|MatchHeader|SanitizeImage``
