+++ niworkflows/func/util.py
@@ -15,9 +15,8 @@ from ..interfaces.fixes import (
     FixHeaderApplyTransforms as ApplyTransforms,
     FixN4BiasFieldCorrection as N4BiasFieldCorrection,
 )
-from ..interfaces.images import ValidateImage, MatchHeader
+from ..interfaces.header import CopyXForm, ValidateImage, MatchHeader
 from ..interfaces.reportlets.masks import SimpleShowMaskRPT
-from ..interfaces.utils import CopyXForm
 from ..utils.connections import listify
 from ..utils.misc import pass_dummy_scans as _pass_dummy_scans
